+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Compatibility:
 * Implemented `FrozenError#receiver`.
 * `Proc#<<` and `Proc#>>` raises TypeError if passed not callable object (#2004, @ssnickolay).
 * Support time and date related messages for `Time` (#2166).
+* Updated `Dir.{glob,[]}` to raise `ArgumentError` for nul-separated strings.
 
 Performance: